ファイル修正

+++ database/migrations/2022_03_17_175125_create_comments_table.php
@@ -15,7 +15,7 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('article_id');
+            $table->unsignedBigInteger('article_id');
             $table->string('comment', 50);
             $table->timestamps();
             $table->foreign('article_id')
